Updated model schemas and move uuid verification to security util
